Extract SimplifiedRegexCharacterClass into a separate file
