+++ eslint-bridge/src/rules/duplicates-in-character-class.ts
@@ -21,7 +21,8 @@
 
 import { Rule } from 'eslint';
 import { CharacterClass, Flags, Node, RegExpLiteral } from 'regexpp/ast';
-import { getRegexpLocation, SimplifiedRegexCharacterClass, toEncodedMessage } from '../utils';
+import { getRegexpLocation, toEncodedMessage } from '../utils';
+import { SimplifiedRegexCharacterClass } from '../utils/simplified-regex-character-class';
 import { createRegExpRule } from './regex-rule-template';
 
 export const rule: Rule.RuleModule = createRegExpRule(
